requestmock code moved to models
