+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 0.9.1 / ####-##-##
 ==================
 
+  * Webworker module and assertions
   * Logo link fixed in readme
-  * Webworker module
 
 0.9.0 / 2018-06-24
 ==================
