+++ lib/utils.js
@@ -1,4 +1,7 @@
 "use strict";
+const querystring = require('querystring');
+const URL = require('url').URL;
+
 
 module.exports = {
     isNumber(n) {
@@ -47,5 +50,14 @@ module.exports = {
             if(obj1[k] !== obj2[k]) return false;
         }
         return true;
+    },
+    parseQueryString(qs) {
+        if(typeof qs === 'string') {
+            return querystring.parse(qs);
+        } else if(qs instanceof URL) {
+            qs = qs.searchParams.toString();
+            return querystring.parse(qs);
+        } else return qs;
     }
+
 };
